Delete Quiz

+++ frontend/src/app/pages/admin/view-quizzes/view-quizzes.component.ts
@@ -26,16 +26,26 @@ export class ViewQuizzesComponent implements OnInit {
   }
 
   deleteQuiz(quizId: any){
-    this._quiz.deleteQuiz(quizId).subscribe(
-    (data)=>{
-      this.quizzes = this.quizzes.filter((quiz)=>quiz['qid'] != quizId);
-      Swal.fire('Success', 'Quiz Deleted!', 'success');
-    },
-    (error)=>{
-      console.log(error)
-      Swal.fire('Error!', 'Error in deleting quiz!' ,'error');
-    }
-      
+    Swal.fire({
+      icon: 'warning',
+      title: 'Are you sure?',
+      confirmButtonText: 'Delete',
+      showCancelButton: true
+    }).then(
+      (result)=>{
+        if(result.isConfirmed){
+          this._quiz.deleteQuiz(quizId).subscribe(
+            (data)=>{
+              this.quizzes = this.quizzes.filter((quiz)=>quiz['qid'] != quizId);
+              Swal.fire('Success', 'Quiz Deleted!', 'success');
+            },
+            (error)=>{
+              console.log(error)
+              Swal.fire('Error!', 'Error in deleting quiz!' ,'error');
+            } 
+          )
+        }
+      }
     )
   }
